Second commit

+++ src/main/java/org/example/modelo/Funcionario.java
@@ -1,7 +1,9 @@
 package org.example.modelo;
 
+import java.math.BigDecimal;
+
 public class Funcionario extends FuncionarioBase {
-    public Funcionario(String nome, String endereco, double salario, Setor setor, Cargo cargo) {
+    public Funcionario(String nome, String endereco, BigDecimal salario, Setor setor, Cargo cargo) {
         super(nome, endereco, salario, setor, cargo);
     }
 }
