+++ src/main/java/org/example/modelo/FuncionarioBase.java
@@ -3,6 +3,7 @@ package org.example.modelo;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -12,11 +13,11 @@ public abstract class FuncionarioBase {
     private String nome;
     private List<String> telefones;
     private String endereco;
-    private double salario;
+    private BigDecimal salario;
     private Setor setor;
     private Cargo cargo;
 
-    public FuncionarioBase(String nome, String endereco, double salario, Setor setor, Cargo cargo) {
+    public FuncionarioBase(String nome, String endereco, BigDecimal salario, Setor setor, Cargo cargo) {
         this.nome = nome;
         this.endereco = endereco;
         this.salario = salario;
