+++ src/main/java/org/example/modelo/FuncionarioTerceirizado.java
@@ -3,13 +3,15 @@ package org.example.modelo;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class FuncionarioTerceirizado extends FuncionarioBase {
     private String empresaContratada;
     private int tempoPrevistoPermanencia;
 
-    public FuncionarioTerceirizado(String nome, String endereco, double salario, Setor setor, Cargo cargo, String empresaContratada, int tempoPrevistoPermanencia) {
+    public FuncionarioTerceirizado(String nome, String endereco, BigDecimal salario, Setor setor, Cargo cargo, String empresaContratada, int tempoPrevistoPermanencia) {
         super(nome, endereco, salario, setor, cargo);
         this.empresaContratada = empresaContratada;
         this.tempoPrevistoPermanencia = tempoPrevistoPermanencia;
